Backend/requirements.txt
bcrypt==4.0.1
asgiref==3.8.1
channels==4.1.0
channels-redis==4.2.0
colorama==0.4.6
crispy-bootstrap5==2024.2  # Note the unique version
daphne==4.0.0
dj-database-url==2.1.0
Django==5.0.3
django-channels==0.7.0
django-cors-headers==4.3.1
django-environ==0.9.0
django-environ==0.9.0  # Consider using the latest version (0.11.2)
django-googledrive-storage==1.6.0
django-rest-framework==0.1.0
django-rest-framework==0.1.0  # Consider using the latest version (3.15.1)
Django-Verify-Email==2.0.3
django-widget-tweaks==1.4.8
django-widget-tweaks==1.4.8  # Consider using the latest version (1.5.0)
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf_yasg==1.21.7
email_validator==2.1.1
eventlet==0.34.2
executing==1.2.0
gunicorn==21.2.0
iniconfig==2.0.0
openai==1.17.0
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python_jose==3.3.0
python-jose==3.3.0
python-socketio==5.11.1
uvicorn==0.29.0
Werkzeug==2.3.8
websocket==0.2.1
websocket-client==1.6.4
Werkzeug==2.3.8
whitenoise==6.6.0
